requirements.txt
python-telegram-bot==21.9
gspread==6.1.4
oauth2client==4.1.3
APScheduler==3.10.4
pytz==2024.2
python-dotenv==1.0.1
flask==3.0.0
schedule==1.2.0
